+++ flutter_local_notifications/lib/src/platform_specifics/android/method_channel_mappers.dart
@@ -249,7 +249,7 @@ extension AndroidNotificationDetailsMapper on AndroidNotificationDetails {
         'fullScreenIntent': fullScreenIntent,
         'shortcutId': shortcutId,
         'additionalFlags': additionalFlags,
-        'customLayoutLegacyName': customLayoutLegacyName,
+        'customLayoutLegacyName': customLayoutLegacyName ?? customLayoutName,
         'customLayoutCollapsedName': customLayoutCollapsedName,
         'customLayoutExpandedName': customLayoutExpandedName,
         'subText': subText,